Add files via upload
Introduction:
Hotel booking cancellations can be costly for hotels and frustrating for guests. Analyzing data on cancellations can help hotels reduce cancellation rates and improve profitability. This report presents the results of the exploratory data analysis phase of a hotel booking cancellation data analysis project. We describe the data sources and preprocessing steps, as well as the analysis techniques used to explore the data. The goal is to identify patterns and trends and generate insights that can inform subsequent phases of the project.